% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fpkm.R
\name{calculate_fpkm}
\alias{calculate_fpkm}
\title{Calculate FPKM Using For Loop}
\usage{
calculate_fpkm(data, count_cols)
}
\arguments{
\item{data}{A data.frame containing RNA-seq data with gene lengths and counts.}

\item{count_cols}{A character vector specifying the column names that contain the raw count data.}
}
\value{
A matrix of FPKM values with gene names as rownames and sample names as colnames.
}
\description{
This function calculates FPKM values for each gene in each sample using a for loop.
}
