% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform_fpkm.R
\name{log_transform_fpkm}
\alias{log_transform_fpkm}
\title{Log-Transform FPKM Values}
\usage{
log_transform_fpkm(fpkm_data, pseudocount = 1)
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}

\item{pseudocount}{A pseudocount to avoid log(0), default is 1.}
}
\value{
A data frame of log-transformed FPKM values.
}
\description{
Applies a log2 transformation to FPKM values.
}
