% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fpkm_tables.R
\name{merge_fpkm_tables}
\alias{merge_fpkm_tables}
\title{Merge FPKM Tables}
\usage{
merge_fpkm_tables(fpkm_list)
}
\arguments{
\item{fpkm_list}{A list of data frames of FPKM values to merge}
}
\value{
A merged data frame
}
\description{
Merge FPKM Tables
}
