% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{normalize_counts_tpm}
\alias{normalize_counts_tpm}
\title{Normalize Counts to TPM}
\usage{
normalize_counts_tpm(counts, lengths)
}
\arguments{
\item{counts}{A vector of gene counts.}

\item{lengths}{A vector of gene lengths.}
}
\value{
A vector of TPM values.
}
\description{
Normalize gene expression data (counts) to Transcripts Per Million (TPM).
}
