% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_heatmap.R
\name{plot_expression_heatmap}
\alias{plot_expression_heatmap}
\title{Plot Expression Heatmap}
\usage{
plot_expression_heatmap(fpkm_data, top_n = 20)
}
\arguments{
\item{fpkm_data}{A data frame or matrix where rows represent genes and columns represent samples.}

\item{top_n}{An integer specifying the number of top genes (by variance) to display (default 20).}
}
\value{
A heatmap plot of the scaled expression values.
}
\description{
This function takes a matrix or data frame of FPKM values and visualizes a heatmap of the most variable genes.
}
