% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpkm_distribution.R
\name{plot_fpkm_distribution}
\alias{plot_fpkm_distribution}
\title{Plot FPKM Distribution}
\usage{
plot_fpkm_distribution(fpkm_values)
}
\arguments{
\item{fpkm_values}{Numeric vector of FPKM values}
}
\value{
A ggplot histogram
}
\description{
Plot FPKM Distribution
}
