% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_expression.R
\name{plot_gene_expression}
\alias{plot_gene_expression}
\title{Plot Gene Expression Over Samples}
\usage{
plot_gene_expression(fpkm_data, gene, plot_type = "bar")
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}

\item{gene}{The gene of interest.}

\item{plot_type}{The type of plot ("bar" or "line").}
}
\value{
A plot of the gene expression.
}
\description{
Plots the expression of a specific gene across different samples.
}
