% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fpkm_data.R
\name{process_fpkm_data}
\alias{process_fpkm_data}
\title{Process FPKM Data Frame}
\usage{
process_fpkm_data(data)
}
\arguments{
\item{data}{Data frame with \code{Counts} and \code{Length_kb} columns}
}
\value{
Data frame with added \code{FPKM} column, sorted by FPKM
}
\description{
Process FPKM Data Frame
}
