% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_expression}
\alias{summarize_expression}
\title{Summarize Expression Data}
\usage{
summarize_expression(fpkm_data)
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}
}
\value{
A data frame with the summary statistics for each gene.
}
\description{
Summarizes gene expression data with mean, median, and variance.
}
