% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fpkm.R
\name{calculate_fpkm}
\alias{calculate_fpkm}
\title{Calculate FPKM}
\usage{
calculate_fpkm(counts, lengths)
}
\arguments{
\item{counts}{Numeric vector of read counts}

\item{lengths}{Numeric vector of gene lengths (in kilobases)}
}
\value{
Numeric vector of FPKM values
}
\description{
Calculate FPKM
}
