% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_top_expressed_genes}
\alias{get_top_expressed_genes}
\title{Get Top Expressed Genes}
\usage{
get_top_expressed_genes(fpkm_data, top_n = 10)
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}

\item{top_n}{The number of top expressed genes to return.}
}
\value{
A data frame of the top expressed genes.
}
\description{
Retrieves the top N most highly expressed genes.
}
