% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_expressed_genes.R
\name{get_top_expressed_genes}
\alias{get_top_expressed_genes}
\title{Get Top Expressed Genes}
\usage{
get_top_expressed_genes(fpkm_data, top_n = 50)
}
\arguments{
\item{fpkm_data}{A data frame or matrix of (optionally log-transformed) FPKM values, where rows are genes and columns are samples.}

\item{top_n}{The number of top expressed genes to return (default is 50).}
}
\value{
A character vector of gene names corresponding to the top expressed genes.
}
\description{
Retrieves the top N most highly expressed genes based on average expression across samples.
}
