% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_heatmap.R
\name{plot_expression_heatmap}
\alias{plot_expression_heatmap}
\title{Plot Expression Heatmap for Top N Genes}
\usage{
plot_expression_heatmap(fpkm_data, top_n = 10)
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}

\item{top_n}{The number of top expressed genes to plot.}
}
\value{
A heatmap plot.
}
\description{
Generates a heatmap for the top N expressed genes.
}
