% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_heatmap.R
\name{plot_expression_heatmap}
\alias{plot_expression_heatmap}
\title{Plot Expression Heatmap}
\usage{
plot_expression_heatmap(fpkm_data)
}
\arguments{
\item{fpkm_data}{A data frame or matrix of FPKM values (genes as rows, samples as columns).}
}
\value{
A pheatmap object.
}
\description{
Plot Expression Heatmap
}
