% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpkm_boxplot.R
\name{plot_fpkm_boxplot}
\alias{plot_fpkm_boxplot}
\title{Plot FPKM Boxplot}
\usage{
plot_fpkm_boxplot(fpkm_data)
}
\arguments{
\item{fpkm_data}{A data frame or matrix of FPKM values (genes x samples).}
}
\value{
A ggplot2 boxplot object.
}
\description{
This function takes a wide-format FPKM matrix and plots the distribution per sample.
}
