% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpkm_boxplot.R
\name{plot_fpkm_boxplot}
\alias{plot_fpkm_boxplot}
\title{Plot FPKM Distribution Boxplot}
\usage{
plot_fpkm_boxplot(log_fpkm)
}
\arguments{
\item{log_fpkm}{A data frame of log-transformed FPKM values.}
}
\value{
A ggplot2 boxplot.
}
\description{
Generates a boxplot of FPKM values across samples.
}
