% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_pca_fpkm}
\alias{plot_pca_fpkm}
\title{Plot PCA of FPKM Values}
\usage{
plot_pca_fpkm(fpkm_data)
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}
}
\value{
A PCA plot.
}
\description{
Generates a PCA plot based on FPKM values for all genes.
}
