% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_fpkm.R
\name{plot_pca_fpkm}
\alias{plot_pca_fpkm}
\title{PCA Plot of FPKM Matrix}
\usage{
plot_pca_fpkm(fpkm_data)
}
\arguments{
\item{fpkm_data}{A numeric matrix or data frame with genes as rows and samples as columns.}
}
\value{
A ggplot2 PCA plot object.
}
\description{
This function performs PCA on a log2-transformed FPKM matrix.
}
