% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_correlation.R
\name{plot_sample_correlation}
\alias{plot_sample_correlation}
\title{Plot Sample Correlation Heatmap}
\usage{
plot_sample_correlation(numeric_data)
}
\arguments{
\item{numeric_data}{A data frame or matrix of numeric FPKM values (genes as rows, samples as columns).}
}
\value{
A pheatmap object.
}
\description{
Plot Sample Correlation Heatmap
}
