% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_sample_correlation}
\alias{plot_sample_correlation}
\title{Plot Sample Correlation Heatmap}
\usage{
plot_sample_correlation(fpkm_data)
}
\arguments{
\item{fpkm_data}{A data frame of FPKM values.}
}
\value{
A heatmap plot.
}
\description{
Generates a heatmap showing the correlation between samples based on FPKM values.
}
