% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_seq_data.R
\name{read_rna_seq_data}
\alias{read_rna_seq_data}
\title{Read RNA-seq Data File}
\usage{
read_rna_seq_data(file_path)
}
\arguments{
\item{file_path}{Path to RNA-seq data file}
}
\value{
A data frame with GeneID, Counts, and Length_kb
}
\description{
Read RNA-seq Data File
}
